Moved analysis functions to new module
